cập nhật code lần thứ 98

+++ user/PHP_Function/xoa_dia_chi.php
@@ -1,8 +1,13 @@
 <?php require_once('../../query.php');
+
 $id = $_POST["id_address"];
-$sql = execute("delete from diachikhach where id_addr = {$id}");
-if ($sql) {
-    echo "Xóa địa chỉ thành công !";
-} else {
-    echo "em dep lam";
+$check= executeSingleResult("select * from diachikhach where id_addr = '{$id}'");
+//die("select * from diachikhach where id_addr = $id");
+if($check['loai']==1){
+    echo 'fail';
+    die();
+}else {
+    $sql = execute("delete from diachikhach where id_addr = {$id}");
+    echo 'success';
 }
+
